grpc-python/pyproject.toml
[tool.poetry]
name = "grpc-python"
version = "0.1.0"
description = ""
authors = ["fhassis <fhassis2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.48.1"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.48.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
